KnowledgeGraphClass/TP5/app/pyproject.toml
[project]
name = "tp5-knowledge-graph-rag"
version = "1.0.0"
description = "TP5 - Building Knowledge Graphs from Unstructured Text using LLMs, Neo4j, and LangChain"
requires-python = ">=3.11"
dependencies = [
    # Neo4j connectivity
    "neo4j>=5.20.0",

    # Data manipulation
    "pandas>=2.1.0",
    "numpy>=1.24.0",

    # Visualization
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "networkx>=3.1",
    "pyvis>=0.3.2",

    # Jupyter
    "jupyterlab>=4.0.0",
    "ipykernel>=6.25.0",
    "ipywidgets>=8.1.0",

    # LangChain ecosystem
    "langchain>=0.1.0",
    "langchain-community>=0.0.10",
    "langchain-experimental>=0.0.40",
    "langchain-neo4j>=0.0.3",
    "langchain-ollama>=0.1.0",
    "langchain-text-splitters>=0.0.1",

    # Utilities
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "wikipedia-api>=0.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
